requirements.txt
python-dotenv
streamlit
watchdog
haystack-ai
requests